Implementation Attempt
I tried to rig up the climbing system with the assets Invector gave me but the climbable objects prefabs didn't come with the box colliders and scripts that allow for climbing...I think. I couldn't find the script for TriggeringGenericAction and the prefabs didn't come with them, so I think at this point I'm better off coding it based on tutorials/learning from open source software.